requirements.txt
requests
beautifulsoup4
lxml
google-generativeai
python-dotenv
tiktoken
sentence-transformers
chromadb
EbookLib
tqdm